x/component: refactor into more realistic architecture
This example has grown in complexity to the point where having
all of the code in a single file is no longer helpful. This commit
introduces the abstractions necessary to split this app into
several files and packages in a way that is realistic for how
people actually build applications with Gio.

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>