7gui/counter: update counter before displaying it
Opening the app and "clicking" the button with say <Tab><Space>
does not increment the counter prior to this change.
Fix this by first processing the clicks, then render.

This doesn't normally happen if you use the mouse, as the animation
and / or the cursor handling invalidate the ui passively.

Signed-off-by: Reto Brunner <reto@slightlybroken.com>